BLD: revise all build scripts to make them work with libc++
When calling `linkLibCpp()` in build.zig, Zig compiler would generate
`libc++.a` and `libc++abi.a` and make the target link aginst them.

But since it's `libc++`, we might ran into compatibility issue of C++
ABI when the toy dialect library is not built with the same stdlib.
(this issue is not discovered until we started working on Ch3)

Though we can still try to link our Zig program to `libstd++`, we cannot
expect there would be no further issue related to ABI compatibility.

The other approach is trying to build all required libraries with Zig,
but we failed to build LLVM/MLIR with it since some linker flags are not
supported yet (e.g., `-rpath-link`).

So we tried to rebuild LLVM/MLIR with libc++ with clang in 2 stages (see
also `utils/build_llvm_mlir.sh` for further details), then rebuild toy
dialect library and link the Zig program against Zig's `libc++`. And it
did work for now.

related issue: #1