MAINT: replace symlinks of parser related files with actual files in Ch7
In Ch7, we are going to deal with the new type `Struct` in toy dialect.
So that we need to modify "ast/lexer/parser.zig" accordingly.

To make it easier to track the changes when comparing with the orignal
parser files, we made this commit first before applying other necessary
changes.