BUG: pass a buffer for printing token from caller to `lexer.Token.str()`
In the original implementation of `lexer.Token.str()`, it's trying to
print the value to a local buffer. So that the returned slice would
reference to a stack variable which would be invalid after leaving the
function, and this is why we cannot get the right output when printing
that error message.