MAINT: rewrite the approach to get output from command for comparison
In "compare_output.sh", the original approach for getting output from
command would fail after Ch5 (include). Output is still visibile in
terminal if "toyc-chX" is run directly, but it would not be captured
in the neovim's diff view via bash's process substitution.

So we change the approach to capture command output into redirecting
both STDOUR and STDERR to a temporary file for now.